Messenger/requirements.txt
asgiref==3.5.2
auth==0.5.3
automium-web==0.1.1
automium==0.2.6
backports.zoneinfo==0.2.1
blinker==1.5
certifi==2022.9.24
cffi==1.15.1
chainer==7.8.1
channels==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
django-appconf==1.0.5
django-auth==0.1.9
django-autoslug==1.9.8
django-imagekit==4.1.0
Django==4.1.2
djangorestframework==3.14.0
dnspython==2.2.1
eventlet==0.33.1
falcon==3.1.0
filelock==3.8.0
Flask==2.2.2
gevent==22.10.1
greenlet==1.1.3.post0
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
legacy==0.1.6
MarkupSafe==2.1.1
mongoengine==0.24.2
numpy==1.23.4
path==16.5.0
pilkit==2.0
Pillow==9.2.0
pip==21.1.2
protobuf==4.21.8
pycparser==2.21
pymongo==4.3.2
pypandoc==1.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
pyzmq==24.0.1
requests==2.28.1
room==0.1.11
setuptools==57.0.0
six==1.16.0
sqlparse==0.4.3
tornado==6.2
typing-extensions==4.4.0
tzdata==2022.5
urllib3==1.26.12
views==0.3
websocket==0.2.1
Werkzeug==2.2.2
wheel==0.36.2
zipp==3.10.0
zope.event==4.5.0
zope.interface==5.5.0
